I moved the file weatherReport.js inside the routes because it's most closely realted to it and also not used in any other places
